--- a/packages/storage/package.json
+++ b/packages/storage/package.json
@@ -62,7 +62,8 @@
             "esnext.asynciterable",
             "es2017.object"
           ],
-          "allowJs": true
+          "allowJs": true,
+					"target": "ES6"
         }
       }
     },