--- a/packages/storage/src/providers/AWSS3ProviderManagedUpload.ts
+++ b/packages/storage/src/providers/AWSS3ProviderManagedUpload.ts
@@ -24,14 +24,11 @@ import {
 	UploadPartCommand,
 	CompleteMultipartUploadCommand,
 	CompleteMultipartUploadCommandInput,
-	UploadPartCommandOutput,
-	UploadPartCommandInput,
 	ListPartsCommand,
 	AbortMultipartUploadCommand,
 } from '@aws-sdk/client-s3';
 import { AxiosHttpHandler, SEND_PROGRESS_EVENT } from './axios-http-handler';
 import * as events from 'events';
-import { streamCollector } from '@aws-sdk/fetch-http-handler';
 
 const logger = new Logger('AWSS3ProviderManagedUpload');
 
@@ -39,11 +36,17 @@ const localTestingStorageEndpoint = 'http://localhost:20005';
 
 const SET_CONTENT_LENGTH_HEADER = 'contentLengthMiddleware';
 export declare interface Part {
-	bodyPart: any;
+	bodyPart: Blob;
 	partNumber: number;
-	emitter: any;
+	emitter: events.EventEmitter;
 	etag?: string;
 	_lastUploadedBytes: number;
+}
+
+class RNBlob extends Blob {
+	get [Symbol.toStringTag]() {
+		return 'Blob';
+	}
 }
 
 export class AWSS3ProviderManagedUpload {
@@ -63,7 +66,7 @@ export class AWSS3ProviderManagedUpload {
 	private totalBytesToUpload = 0;
 	private emitter = null;
 
-	constructor(params, opts, emitter) {
+	constructor(params, opts, emitter: events.EventEmitter) {
 		this.params = params;
 		this.opts = opts;
 		this.emitter = emitter;
@@ -86,6 +89,8 @@ export class AWSS3ProviderManagedUpload {
 			const numberOfPartsToUpload = Math.ceil(
 				this.totalBytesToUpload / this.minPartSize
 			);
+
+			const parts: Part[] = this.createParts();
 			for (
 				let start = 0;
 				start < numberOfPartsToUpload;
@@ -97,8 +102,10 @@ export class AWSS3ProviderManagedUpload {
 				await this.checkIfUploadCancelled(uploadId);
 
 				// Upload as many as `queueSize` parts simultaneously
-				const parts: Part[] = this.createParts(start);
-				await this.uploadParts(uploadId, parts);
+				await this.uploadParts(
+					uploadId,
+					parts.slice(start, start + this.queueSize)
+				);
 
 				/** Call cleanup a second time in case there were part upload requests
 				 *  in flight. This is to ensure that all parts are cleaned up.
@@ -111,14 +118,10 @@ export class AWSS3ProviderManagedUpload {
 		}
 	}
 
-	private createParts(startPartNumber: number): Part[] {
+	private createParts(): Part[] {
 		const parts: Part[] = [];
-		let partNumber = startPartNumber;
-		for (
-			let bodyStart = startPartNumber * this.minPartSize;
-			bodyStart < this.totalBytesToUpload && parts.length < this.queueSize;
-
-		) {
+		let partNumber = 0;
+		for (let bodyStart = 0; bodyStart < this.totalBytesToUpload; ) {
 			const bodyEnd = Math.min(
 				bodyStart + this.minPartSize,
 				this.totalBytesToUpload
@@ -165,28 +168,34 @@ export class AWSS3ProviderManagedUpload {
 		return response.UploadId;
 	}
 
+	private transformBlob(blob: Blob): RNBlob {
+		return new RNBlob([blob]);
+	}
+
 	/**
 	 * @private Not to be extended outside of tests
 	 * @VisibleFotTesting
 	 */
 	protected async uploadParts(uploadId: string, parts: Part[]) {
-		const promises: Array<Promise<UploadPartCommandOutput>> = [];
-		for (const part of parts) {
-			this.setupEventListener(part);
-			const uploadPartCommandInput: UploadPartCommandInput = {
-				PartNumber: part.partNumber,
-				Body: part.bodyPart,
-				UploadId: uploadId,
-				Key: this.params.Key,
-				Bucket: this.params.Bucket,
-			};
-			const uploadPartCommand = new UploadPartCommand(uploadPartCommandInput);
-			const s3 = await this._createNewS3Client(this.opts, part.emitter);
-			promises.push(s3.send(uploadPartCommand));
-		}
+		console.log('Uploading parts');
 		try {
-			const allResults: Array<UploadPartCommandOutput> = await Promise.all(
-				promises
+			const allResults = await Promise.all(
+				parts.map(async part => {
+					console.log(`Uploading part ${part.partNumber}`);
+					this.setupEventListener(part);
+					const s3 = await this._createNewS3Client(this.opts, part.emitter);
+					return s3.send(
+						new UploadPartCommand({
+							PartNumber: part.partNumber,
+							Body: Platform.isReactNative
+								? this.transformBlob(part.bodyPart)
+								: part.bodyPart,
+							UploadId: uploadId,
+							Key: this.params.Key,
+							Bucket: this.params.Bucket,
+						})
+					);
+				})
 			);
 			// The order of resolved promises is the same as input promise order.
 			for (let i = 0; i < allResults.length; i++) {
@@ -195,6 +204,9 @@ export class AWSS3ProviderManagedUpload {
 					ETag: allResults[i].ETag,
 				});
 			}
+			parts.forEach(part => {
+				this.removeEventListener(part);
+			});
 		} catch (error) {
 			logger.error(
 				'error happened while uploading a part. Cancelling the multipart upload',
@@ -267,6 +279,10 @@ export class AWSS3ProviderManagedUpload {
 		}
 	}
 
+	private removeEventListener(part: Part) {
+		part.emitter.removeAllListeners(SEND_PROGRESS_EVENT);
+	}
+
 	private setupEventListener(part: Part) {
 		part.emitter.on(SEND_PROGRESS_EVENT, progress => {
 			this.progressChanged(
@@ -308,14 +324,6 @@ export class AWSS3ProviderManagedUpload {
 		if (this.isGenericObject(body)) {
 			// Any javascript object
 			return JSON.stringify(body);
-		} else if (this.isBlob(body)) {
-			// If it's a blob, we need to convert it to an array buffer as axios has issues
-			// with correctly identifying blobs in *react native* environment. For more
-			// details see https://github.com/aws-amplify/amplify-js/issues/5311
-			if (Platform.isReactNative) {
-				return await streamCollector(body);
-			}
-			return body;
 		} else {
 			// Files, arrayBuffer etc
 			return body;
@@ -329,11 +337,11 @@ export class AWSS3ProviderManagedUpload {
 		} */
 	}
 
-	private isBlob(body: any) {
+	private isBlob(body: any): body is Blob {
 		return typeof Blob !== 'undefined' && body instanceof Blob;
 	}
 
-	private isGenericObject(body: any) {
+	private isGenericObject(body: any): body is Object {
 		if (body !== null && typeof body === 'object') {
 			try {
 				return !(this.byteLength(body) >= 0);
@@ -350,7 +358,7 @@ export class AWSS3ProviderManagedUpload {
 	 * @private
 	 * creates an S3 client with new V3 aws sdk
 	 */
-	protected async _createNewS3Client(config, emitter?) {
+	protected async _createNewS3Client(config, emitter?: events.EventEmitter) {
 		const credentials = await this._getCredentials();
 		const {
 			region,