--- a/packages/storage/src/providers/axios-http-handler.ts
+++ b/packages/storage/src/providers/axios-http-handler.ts
@@ -133,6 +133,11 @@ export class AxiosHttpHandler implements HttpHandler {
 
 		// From gamma release, aws-sdk now expects all response type to be of blob or streams
 		axiosRequest.responseType = 'blob';
+		
+		// In Axios, Blobs are identified by calling Object.prototype.toString on the object. However, on React Native,
+		// calling Object.prototype.toString on a Blob returns '[object Object]' instead of '[object Blob]', which causes 
+		// Axios to treat Blobs as generic Javascript objects. Therefore we need a to use a custom request transformer
+		// to correctly handle Blob in React Native.
 		if (Platform.isReactNative) {
 			axiosRequest.transformRequest = reactNativeRequestTransformer;
 		}