#### Description of changes
This fixes the memory usage issue for React Native.

The issue was with `streamCollector`. It will load everything from the file into memory before uploading, which will cause an out of memory exception if the file is large enough, slows down upload a lot and it also blocks the UI in the process.

It stems from the problem where Axios identifies the type of data using `Object.prototype.toString`, see: [utils#L147](https://github.com/axios/axios/blob/3958e9fbc8ebc0f72a8ce568cbf86f391d6869f3/lib/utils.js#L147) and  On React Native, calling `Object.prototype.toString` on Blob returns `'[object Object]'` instead of `'[object Blob]'`, it will then call a `JSON.stringify` on it instead of handling a Blob correctly. So we need to use a custom request transformer for React Native to correctly identify a Blob.

see: [defaultTransformer on Axios](https://github.com/axios/axios/blob/3958e9fbc8ebc0f72a8ce568cbf86f391d6869f3/lib/defaults.js#L39)

`createParts` is modified to partition the entire file instead of partition it on the fly due to a GC issue. In React Native Android, when slice of the blob gets GCed, it GCs the entire blob as well, so we need to partition the entire file before actually uploading.

Also removed some `any`s as I go through the code.

I did a simple benchmark with the original implementation versus this new implementation:
Upload speed is 92Mbps

| size | old | new |
| -- | -- | -- |
| 10mb | 43s | 1.6s |
| 20mb | 76s | 2.4s |
| 50mb | 212s | 5.4s |
| 75mb | ERROR | 8.1s |

Did not test anything beyond 75mb as the old implementation errors out on anything 75MB+

**TLDR: there should be a significant improvement in upload speed & UI won't be blocking for React Native users.**

Demo (20MB file):

https://user-images.githubusercontent.com/40404256/119397524-18f33480-bc8b-11eb-91ca-bfa14bebda98.mov

**(The demo downloads then upload, but it only calculates the time elapsed for the upload)**

#### Issue #, if available
fixes #6419
fixes #6888

#### Description of how you validated changes
Added unit tests
Manually tested on React Native android & ios

#### Checklist
- [x] PR description included
- [x] `yarn test` passes
- [x] Tests are [changed or added](https://github.com/aws-amplify/amplify-js/blob/main/CONTRIBUTING.md#steps-towards-contributions)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.