@jamesaucode 2021-05-25
> @jamesaucode and @amhinson, do you have an ETA for the final merge of this PR? I am currently blocked on a project because of this issue and would prefer to use this final solution instead of using one of the current workarounds.
> Thanks in advance!

Hey @novakzaballa, the changes are now merged it should be available in our "unstable" branch now