@novakzaballa 2021-05-27
@jamesaucode , After some testing using 4.0.2-unstable.21, I still have out of memory issues uploading large files. Files about 30MB are uploaded larger about 75MB cause app to suddenly close due to out of memory issues. How should I report and document the issue?